remove unecessary operation module
