--- src/operation.rs
@@ -1,4 +1,0 @@
-//! Helpers for DID operations that are independent of DID method.
-
-pub mod document;
-pub mod resolve;
