+++ tests/webvh_resolve.rs
@@ -3,8 +3,10 @@
 use credibil_did::{
     KeyPurpose,
     core::{Kind, OneMany},
-    document::{MethodType, Service, VerificationMethod},
-    operation::document::{Create, DocumentBuilder, Update, VerificationMethodBuilder, VmKeyId},
+    document::{
+        Create, DocumentBuilder, MethodType, Service, Update, VerificationMethod,
+        VerificationMethodBuilder, VmKeyId,
+    },
     webvh::{
         SCID_PLACEHOLDER, Witness, WitnessEntry, WitnessWeight, create::CreateBuilder,
         deactivate::DeactivateBuilder, resolve::resolve_log, update::UpdateBuilder,
