+++ tests/webvh_verify.rs
@@ -3,8 +3,10 @@
 use credibil_did::{
     KeyPurpose,
     core::{Kind, OneMany},
-    document::{MethodType, Service, VerificationMethod},
-    operation::document::{Create, DocumentBuilder, VerificationMethodBuilder, VmKeyId},
+    document::{
+        Create, DocumentBuilder, MethodType, Service, VerificationMethod,
+        VerificationMethodBuilder, VmKeyId,
+    },
     webvh::{
         SCID_PLACEHOLDER, Witness, WitnessWeight, create::CreateBuilder, url::default_did,
         verify::verify_proofs,
